Audio & Multithreading Support

+++ engine/src/main/java/dev/gamekit/ui/Constraints.java
@@ -3,8 +3,7 @@ package dev.gamekit.ui;
 import static dev.gamekit.utils.Math.clamp;
 
 /** An immutable record for width and height range values */
-public record Constraints(int minWidth, int maxWidth, int minHeight,
-                          int maxHeight) {
+public record Constraints(double minWidth, double maxWidth, double minHeight, double maxHeight) {
   public Constraints {
     minWidth = Math.max(minWidth, 0);
     maxWidth = Math.max(maxWidth, 0);
@@ -12,26 +11,20 @@ public record Constraints(int minWidth, int maxWidth, int minHeight,
     maxHeight = Math.max(maxHeight, 0);
   }
 
-  /**
-   * Constraints the given width to respect {@link #minWidth} and
-   * {@link #maxWidth}
-   */
-  public int constrainWidth(int width) {
+  /** Constraints the given width to respect {@link #minWidth} and {@link #maxWidth} */
+  public double constrainWidth(double width) {
     return clamp(width, minWidth, maxWidth);
   }
 
-  /**
-   * Constraints the given height to respect {@link #minHeight} and
-   * {@link #maxHeight}
-   */
-  public int constrainHeight(int height) {
+  /** Constraints the given height to respect {@link #minHeight} and {@link #maxHeight} */
+  public double constrainHeight(double height) {
     return clamp(height, minHeight, maxHeight);
   }
 
   @Override
   public String toString() {
     return String.format(
-      getClass().getName() + "[%d<=width<=%d,%d<=height<=%d]",
+      getClass().getName() + "[%.2f<=width<=%.2f,%.2f<=height<=%.2f]",
       minWidth, maxWidth, minHeight, maxHeight
     );
   }
